pyproject.toml
[tool.poetry]
name = "mapsy"
version = "0.1.0"
description = ""
authors = ["Felix Horvat <felix.horvat@ocell.io>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/ocell/mapsy"
repository = "https://github.com/ocell/mapsy"


[tool.poetry.dependencies]
python = "^3.10"
shapely = "^2.0.4"
pillow = "^10.4.0"
pycairo = "^1.26.1"
pyproj = "^3.6.1"
mercantile = "^1.2.1"
affine = "^2.4.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.2"
flake8 = "^7.1.1"
mock = "^5.1.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
